refs #5 Put all the mantid fitting utilities and functions in one file.
