+++ fitbenchmarking/fitting/mantid.py
@@ -1,5 +1,5 @@
 """
-Utility functions used in the mantid fitting process.
+Fittng and utility functions for the mantid fitting algorithms.
 """
 # Copyright &copy; 2016 ISIS Rutherford Appleton Laboratory, NScD
 # Oak Ridge National Laboratory & European Spallation Source
@@ -24,9 +24,49 @@ Utility functions used in the mantid fitting process.
 
 from __future__ import (absolute_import, division, print_function)
 
+import time
 import numpy as np
 import mantid.simpleapi as msapi
+
+from utils.logging_setup import logger
 from fitting.plotting import plot_helper
+from utils import fit_misc
+
+
+def fit(prob, wks, function, minimizer='Levenberg-Marquardt',
+        cost_function='Least squares'):
+    """
+    The mantid fit algorithm.
+
+    @param prob :: object holding the problem information
+    @param wks :: workspace holding the problem data
+    @param function :: the fitted function
+    @param minimizer :: the minimizer used in the fitting process
+    @param cost_function :: the type of cost function used in fitting
+
+    @returns :: the status, either success or failure (str), the fit
+                workspace (mantid wks), containing the fitted data,
+                the corresponding points of the fit and the difference
+                between them, the fitted parameters, their errors [arrays]
+                and how much time it took for the fit to finish (float)
+    """
+
+    fit_result, t_start, t_end = None, None, None
+    try:
+        ignore_invalid = get_ignore_invalid(prob, cost_function)
+        t_start = time.clock()
+        fit_result = msapi.Fit(function, wks, Output='ws_fitting_test',
+                               Minimizer=minimizer, CostFunction=cost_function,
+                               IgnoreInvalidData=ignore_invalid,
+                               StartX=prob.start_x, EndX=prob.end_x)
+        t_end = time.clock()
+    except (RuntimeError, ValueError) as err:
+        logger.error("Warning, fit failed. Going on. Error: " + str(err))
+
+    status, fit_wks, fin_function_def, runtime = \
+    parse_result(fit_result, t_start, t_end)
+
+    return status, fit_wks, fin_function_def, runtime
 
 
 def parse_result(fit_result, t_start, t_end):
@@ -118,7 +158,7 @@ def function_definitions(prob):
     return function_defs
 
 
-def ignore_invalid(prob, cost_function):
+def get_ignore_invalid(prob, cost_function):
     """
     Helper function that sets the whether the mantid fitting algorithm
     ignores invalid data or not. This depends on the cost function.
@@ -162,6 +202,32 @@ def parse_nist_function_definitions(prob, nb_start_vals):
                              format(prob.equation, start_val_str))
 
     return function_defs
+
+def chisq(status, fit_wks, min_chi_sq, best_fit, minimizer):
+    """
+    Function that calcuates the chisq obtained through the
+    mantid fitting algorithm and find the best fit out of all
+    the attempted minimizers.
+
+    @param status :: the status of the fit, i.e. success or failure
+    @param fit_wks :: the fit workspace
+    @param min_chi_sq :: the minimium chisq (at the moment)
+    @param best_fit :: the best fit (at the moment)
+    @param minimizer :: minimizer with which the fit_wks was obtained
+
+    @returns :: the chi squared, the new/unaltered minimum chi squared
+                and the new/unaltered best fit data object
+    """
+
+    if status != 'failed':
+        chi_sq = fit_misc.compute_chisq(fit_wks.readY(2))
+        if chi_sq < min_chi_sq and not chi_sq == np.nan:
+            best_fit = optimum(fit_wks, minimizer, best_fit)
+            min_chi_sq = chi_sq
+    else:
+        chi_sq = np.nan
+
+    return chi_sq, min_chi_sq, best_fit
 
 
 def setup_errors(prob):
