+++ fitbenchmarking/fitting/tests/mantid_tests.py
@@ -12,36 +12,18 @@ parent_dir = os.path.dirname(os.path.normpath(test_dir))
 main_dir = os.path.dirname(os.path.normpath(parent_dir))
 sys.path.insert(0, main_dir)
 
-from fitting.mantid_utils import parse_result
-from fitting.mantid_utils import optimum
-from fitting.mantid_utils import wks_cost_function
-from fitting.mantid_utils import function_definitions
-from fitting.mantid_utils import ignore_invalid
-from fitting.mantid_utils import parse_nist_function_definitions
-from fitting.mantid_utils import setup_errors
-
+from fitting.mantid import fit
+from fitting.mantid import parse_result
+from fitting.mantid import optimum
+from fitting.mantid import wks_cost_function
+from fitting.mantid import function_definitions
+from fitting.mantid import get_ignore_invalid
+from fitting.mantid import parse_nist_function_definitions
+from fitting.mantid import setup_errors
 from utils import test_problem
 
 
-class MantidUtilsTests(unittest.TestCase):
-
-    def Neutron_problem(self):
-        """
-        Sets up the problem object for the neutron problem file:
-        ENGINX193749_calibration_peak19.txt
-        """
-
-        prob = test_problem.FittingTestProblem()
-        prob.name = 'ENGINX 193749 calibration, spectrum 651, peak 19'
-        prob.equation = ("name=LinearBackground;"
-                         "name=BackToBackExponential,"
-                         "I=597.076,A=1,B=0.05,X0=24027.5,S=22.9096")
-        prob.starting_values = None
-        prob.start_x = 23919.5789114
-        prob.end_x = 24189.3183142
-
-        return prob
-
+class MantidTests(unittest.TestCase):
 
     def NIST_problem(self):
         """
@@ -75,6 +57,91 @@ class MantidUtilsTests(unittest.TestCase):
         return prob
 
 
+    def Neutron_problem(self):
+        """
+        Sets up the problem object for the neutron problem file:
+        ENGINX193749_calibration_peak19.txt
+        """
+
+        prob = test_problem.FittingTestProblem()
+        prob.name = 'ENGINX 193749 calibration, spectrum 651, peak 19'
+        prob.equation = ("name=LinearBackground;"
+                         "name=BackToBackExponential,"
+                         "I=597.076,A=1,B=0.05,X0=24027.5,S=22.9096")
+        prob.starting_values = None
+        prob.start_x = 23919.5789114
+        prob.end_x = 24189.3183142
+
+        return prob
+
+
+    def setup_problem_Misra1a_success(self):
+        """
+        Helper function.
+        Sets up the parameters needed to run fit_algorithms.mantid
+        """
+
+        prob = self.NIST_problem()
+        wks = msapi.CreateWorkspace(DataX=prob.data_x,
+                                    DataY=prob.data_y,
+                                    DataE=np.sqrt(prob.data_y))
+        function = ("name=UserFunction, Formula=b1*(1-exp(-b2*x)), "
+                    "b1=500.0,b2=0.0001,")
+        minimizer = 'Levenberg-Marquardt'
+        cost_function = 'Least squares'
+
+        return prob, wks, function, minimizer, cost_function
+
+
+    def setup_problem_Misra1a_fail(self):
+        """
+        Helper function.
+        Sets up the parameters needed to run fit_algorithms.mantid
+        but fail due to incorrect minimizer name.
+        """
+
+        prob = self.NIST_problem()
+        wks = msapi.CreateWorkspace(DataX=prob.data_x,
+                                    DataY=prob.data_y,
+                                    DataE=np.sqrt(prob.data_y))
+        function = \
+        "name=UserFunction,Formula=b1*(1-exp(-b2*x)),b1=500.0,b2=0.0001"
+        minimizer = 'Levenberg-Merquardtss'
+        cost_function = 'Least squared'
+
+        return prob, wks, function, minimizer, cost_function
+
+
+    def expected_results_problem_Misra1a_success(self):
+        """
+        Helper function.
+        Sets up the expected results after running
+        fit_algorithms.mantid with Misra1a.dat problem data.
+        """
+
+        fit_status = 'success'
+        fin_function_def = \
+        "name=UserFunction,Formula=b1*(1-exp( -b2*x)),b1=234.534,b2=0.00056228"
+
+        return fit_status, fin_function_def
+
+
+    def expected_results_problem_Misra1a_fail(self):
+        """
+        Helper function.
+        Sets up the expected failure results after running
+        fit_algorithms.mantid with Misra1a.dat problem data but
+        fail parameters.
+        """
+
+        status = 'failed'
+        fit_wks = None
+        fin_function_def = None
+        runtime = np.nan
+
+        return status, fit_wks, fin_function_def, runtime
+
+
     def create_wks_NIST_problem_with_errors(self):
         """
         Helper function.
@@ -99,6 +166,36 @@ class MantidUtilsTests(unittest.TestCase):
         prob = self.NIST_problem()
         wks_exp = msapi.CreateWorkspace(DataX=prob.data_x, DataY=prob.data_y)
         return wks_exp
+
+
+    def test_fitAlgorithms_mantid_return_success_for_NIST_Misra1a_prob_file(self):
+
+        prob, wks, function, minimizer, cost_function = \
+        self.setup_problem_Misra1a_success()
+
+        status, fit_wks, fin_function_def, runtime = \
+        fit(prob, wks, function, minimizer, cost_function)
+        status_expected, fin_function_def_expected = \
+        self.expected_results_problem_Misra1a_success()
+
+        self.assertEqual(status_expected, status)
+        self.assertEqual(fin_function_def_expected[:44], fin_function_def[:44])
+
+
+    def test_runFit_mantidFit_fails(self):
+
+        prob, wks, function, minimizer, cost_function = \
+        self.setup_problem_Misra1a_fail()
+
+        status, fit_wks, fin_function_def, runtime = \
+        fit(prob, wks, function, minimizer, cost_function)
+        (status_expected, fit_wks_expected, fin_function_def_expected,
+        runtime_expected) = self.expected_results_problem_Misra1a_fail()
+
+        self.assertEqual(status_expected, status)
+        self.assertEqual(fin_function_def_expected, fin_function_def)
+        np.testing.assert_equal(runtime_expected, runtime)
+        np.testing.assert_equal(fit_wks_expected, fit_wks)
 
 
     def test_wksCostFunction_return_with_errors(self):
@@ -159,7 +256,7 @@ class MantidUtilsTests(unittest.TestCase):
         prob.name = 'notWish'
         cost_function = 'Least squares'
 
-        ign_invalid = ignore_invalid(prob, cost_function)
+        ign_invalid = get_ignore_invalid(prob, cost_function)
         ign_invalid_expected = True
 
         self.assertEqual(ign_invalid_expected, ign_invalid)
@@ -171,7 +268,7 @@ class MantidUtilsTests(unittest.TestCase):
         prob.name = 'WISH17701lol'
         cost_function = 'Least squares'
 
-        ign_invalid = ignore_invalid(prob, cost_function)
+        ign_invalid = get_ignore_invalid(prob, cost_function)
         ign_invalid_expected = False
 
         self.assertEqual(ign_invalid_expected, ign_invalid)
